requirements.txt
# HubSpot Report Automation Dependencies
requests>=2.31.0
pandas>=2.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
